+++ src/opcodes/Push.cpp
@@ -37,7 +37,7 @@ Instruction* Push::CreateInstruction(Memory::MemoryOffset& memLoc, Processor* pr
 	switch(*opLoc) {
 		case PUSH_MOD16:
 		{
-			if(((*(opLoc + 1) & 0x38) >> 3) != PUSH_SUB_OPCODE)
+			if((unsigned int)((*(opLoc + 1) & 0x38) >> 3) != PUSH_SUB_OPCODE)
 				return newPush;
 			Operand* dst = ModrmOperand::GetModrmOperand(proc, opLoc, ModrmOperand::MOD, 2);
 			snprintf(buf, 65, "PUSH %s", dst->GetDisasm().c_str());
